+++ incl/KrimzLib/math.h
@@ -4,12 +4,12 @@
 namespace kl {
 	namespace math {
 		// Returns x from the given y of the line that goes through points a and b
-		double lineX(kl::vec2 a, kl::vec2 b, double y) {
+		float lineX(kl::vec2 a, kl::vec2 b, float y) {
 			return ((y - a.y) * (b.x - a.x)) / (b.y - a.y) + a.x;
 		}
 
 		// Returns y from the given x of the line that goes through points a and b
-		double lineY(kl::vec2 a, kl::vec2 b, double x) {
+		float lineY(kl::vec2 a, kl::vec2 b, float x) {
 			return ((b.y - a.y) * (x - a.x)) / (b.x - a.x) + a.y;
 		}
